Add User_Aware_Interface_ContractTests trait and basic User_TransientsTest class.
