+++ tests/phpunit/integration/Core/Storage/User_OptionsTest.php
@@ -12,12 +12,15 @@ namespace Google\Site_Kit\Tests\Core\Storage;
 
 use Google\Site_Kit\Context;
 use Google\Site_Kit\Core\Storage\User_Options;
+use Google\Site_Kit\Tests\Core\Storage\User_Aware_Interface_ContractTests;
 use Google\Site_Kit\Tests\TestCase;
 
 /**
  * @group Storage
  */
 class User_OptionsTest extends TestCase {
+
+	use User_Aware_Interface_ContractTests;
 
 	public function test_get() {
 		$user_id = $this->factory()->user->create();
@@ -105,33 +108,6 @@ class User_OptionsTest extends TestCase {
 		$this->assertFalse( metadata_exists( 'user', $user_id, 'test-key' ) );
 	}
 
-	public function test_get_user_id() {
-		$user_id = $this->factory()->user->create();
-		$context = new Context( GOOGLESITEKIT_PLUGIN_MAIN_FILE );
-		$options = new User_Options( $context, $user_id );
-		$this->assertSame( $user_id, $options->get_user_id() );
-	}
-
-	public function test_switch_user() {
-		$user_id_a = $this->factory()->user->create();
-		$user_id_b = $this->factory()->user->create();
-		update_user_option( $user_id_a, 'test-key', 'test-value-a' );
-		update_user_option( $user_id_b, 'test-key', 'test-value-b' );
-
-		// User Options will use the current user if not passed explicitly.
-		wp_set_current_user( $user_id_a );
-
-		$context = new Context( GOOGLESITEKIT_PLUGIN_MAIN_FILE );
-		$options = new User_Options( $context );
-		$this->assertFalse( $context->is_network_mode() );
-
-		$this->assertEquals( 'test-value-a', $options->get( 'test-key' ) );
-
-		$options->switch_user( $user_id_b );
-
-		$this->assertEquals( 'test-value-b', $options->get( 'test-key' ) );
-	}
-
 	protected function network_activate_site_kit() {
 		add_filter(
 			'pre_site_option_active_sitewide_plugins',
@@ -140,4 +116,12 @@ class User_OptionsTest extends TestCase {
 			}
 		);
 	}
+
+	protected function create_user_aware_instance( $user_id ) {
+		return new User_Options(
+			new Context( GOOGLESITEKIT_PLUGIN_MAIN_FILE ),
+			$user_id
+		);
+	}
+
 }
